\name{trapezoidal.integration}
\alias{trapezoidal.integration}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{To compute the peak area under the each fitting peaks by trapezoidal integration.}
\description{
 To compute the peak area under the each fitting peaks by trapezoidal integration.
}
\usage{
trapezoidal.integration(x,yf)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{The values of x-axis}
  \item{yf}{The yf is a collection of each fitting peaks.}
}

\value{
  the fitted vector
}

\author{Zhimin Zhang, Xia Tong}

\keyword{integration}

