\name{peakheights}
\alias{peakheights}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{estimate the peak height by Gaussian fitting}
\description{
	estimate the peak height by Gaussian fitting
}
\usage{
peakheights(positionwidth,x,y)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
	\item{x}{The values of x-axis}
	\item{y}{The values of y-axis}
	\item{positionwidth}{The positionwidth is the results of the estimation or GAs for position or width.}
}

\value{
  the fitted vector
}

\author{Zhimin Zhang, Xia Tong}

\seealso{\code{\link{peakshape}}}

\keyword{peakheights}