\name{peakshape}
\alias{peakshape}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{choose the kinds of peak function}
\description{
	choose the kinds of peak function
}
\usage{
peakshape(x,position,width,shape)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{The values of x-axis}
  \item{position}{The position of the peak}
  \item{width}{The position of the peak}
  \item{shape}{if the shape==1, the peak function is the Gaussian, and if the shape==else, the peak function is the Lorentzian}
}

\value{
  the fitted vector
}

\author{Zhimin Zhang, Xia Tong}

\keyword{peakshape}
