\name{positionwidth}
\alias{positionwidth}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{extract the fitting position and width of peaks by GA}
\description{
	This function can extract the fitting position and width of peaks gy GA.
}
\usage{
	positionwidth(fitresults)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
	\item{fitresults}{the results of GA}
}

\value{
  the fitted vector
}

\author{Zhimin Zhang, Xia Tong}

\keyword{positionwidth}
	
