\name{areayf}
\alias{areayf}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{the shape and area of the fitting peaks}
\description{
 To compute the peak shape and area of each fitting peaks.
}
\usage{
areayf(fitresults,x,peakheight)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{fitresults}{The fitresults are the fitting results of GA. }
  \item{x}{The values of x-axis}
  \item{peakheight}{The peakheight is the results of the PEAKHEIGHTS function.}
}


\value{
  the fitted vector
}


\author{Zhimin Zhang, Xia Tong}

\seealso{\code{\link{peakshape, PEAKHEIGHTS, integration}}}

\keyword{areayf}

