\name{peakdetectionresidual}
\alias{peakdetectionresidual}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{detect the peak position in the residual signal}
\description{
	detect the peak position in the residual signal
}
\usage{
peakdetectionresidual(x,y)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{The values of x-axis}
  \item{y}{The values of y-axis}
}

\value{
  the fitted vector
}

\author{Zhimin Zhang, Xia Tong}

\seealso{\code{\link{cwt}}}

\keyword{peakdetectionresidual}