\name{peakshape}
\alias{peakshape}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{the function of peakshape}
\description{
	the function of peakshape
}
\usage{
	Gaussian(x,position,width)
	lorentzian(x,position,width)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
	\item{x}{The values of x-axis}
    \item{position}{the peak position}
    \item{width}{the peak width}
}

\value{
  the fitted vector
}

\author{Zhimin Zhang, Xia Tong}

\keyword{peakshape}
